Make the changes requested
