+++ src/App.js
@@ -1,33 +1,15 @@
+import React from 'react';
+
 import Calculator from './components/Calculator';
+import buttonsInfo from './buttons';
 
-function App() {
-  const buttonsInfo = [
-    { text: 'AC', changeBg: false },
-    { text: '+/-', changeBg: false },
-    { text: '%', changeBg: false },
-    { text: '/', changeBg: true },
-    { text: '7', changeBg: false },
-    { text: '8', changeBg: false },
-    { text: '9', changeBg: false },
-    { text: 'x', changeBg: true },
-    { text: '4', changeBg: false },
-    { text: '5', changeBg: false },
-    { text: '6', changeBg: false },
-    { text: '-', changeBg: true },
-    { text: '1', changeBg: false },
-    { text: '2', changeBg: false },
-    { text: '3', changeBg: false },
-    { text: '+', changeBg: true },
-    { text: '0', changeBg: false },
-    { text: '.', changeBg: false },
-    { text: '=', changeBg: true },
-  ];
-
-  return (
-    <div className="App">
-      <Calculator buttons={buttonsInfo} />
-    </div>
-  );
+// eslint-disable-next-line react/prefer-stateless-function
+export default class App extends React.Component {
+  render() {
+    return (
+      <div>
+        <Calculator buttons={buttonsInfo} />
+      </div>
+    );
+  }
 }
-
-export default App;
