+++ src/buttons.js
@@ -1,0 +1,22 @@
+const buttonsInfo = [
+  { text: 'AC', changeBg: false },
+  { text: '+/-', changeBg: false },
+  { text: '%', changeBg: false },
+  { text: '÷', changeBg: true },
+  { text: '7', changeBg: false },
+  { text: '8', changeBg: false },
+  { text: '9', changeBg: false },
+  { text: 'x', changeBg: true },
+  { text: '4', changeBg: false },
+  { text: '5', changeBg: false },
+  { text: '6', changeBg: false },
+  { text: '-', changeBg: true },
+  { text: '1', changeBg: false },
+  { text: '2', changeBg: false },
+  { text: '3', changeBg: false },
+  { text: '+', changeBg: true },
+  { text: '0', changeBg: false },
+  { text: '.', changeBg: false },
+  { text: '=', changeBg: true },
+];
+export default buttonsInfo;
